% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculateRange.R
\name{calculateRange}
\alias{calculateRange}
\title{Calculate the range, given the number of observation, for the forecast
#'}
\usage{
calculateRange(obs, n)
}
\arguments{
\item{obs}{Selected number of observation}

\item{n}{Total number of observation}
}
\description{
Calculate the range, given the number of observation, for the forecast
#'
}
\examples{
calculateRange(300,1367)
}
