% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPricePerc.R
\name{getPricePerc}
\alias{getPricePerc}
\title{Get actual Price and Percentage of change of Ethereum Index}
\usage{
getPricePerc()
}
\value{
A class formed by Price and Percentage of change
}
\description{
Get actual Price and Percentage of change of Ethereum Index
}
