% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\docType{class}
\name{pricePerc-class}
\alias{pricePerc-class}
\title{Create a class for get the returned value of the function getPricePerc}
\description{
Create a class for get the returned value of the function getPricePerc
}
