% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOutPricePerc.R
\name{setOutPricePerc}
\alias{setOutPricePerc}
\title{Set the input value in the correct output widget of the UI}
\usage{
setOutPricePerc(price, perc, output)
}
\arguments{
\item{price}{Actual price of Ehtereum index}

\item{perc}{Actual percentage of change in the Ethereum price index}

\item{output}{The output widget of the UI}
}
\description{
Set the input value in the correct output widget of the UI
}
\examples{
setOutPricePerc("$164,37", "(1,23)", output)
}
